% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blog}
\alias{blog}
\title{Blog data}
\format{
A dataset of dimension 280+1:
\describe{
\item{subject}{Anonymized Mechanical Turk Worker ID}
\item{trial}{Trial number, from 1..NNN}
...
}
}
\source{
blogData_train.csv
}
\usage{
blog
}
\description{
Blog data
}
\keyword{datasets}
