% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_CQR_conditional.R
\name{conf_CQR}
\alias{conf_CQR}
\title{Conditional width and coverage for CQR, internal function used inside conf_CQR_conditional}
\usage{
conf_CQR(X1, Y1, X2, Y2, beta, mtry, ntree, alpha = 0.1)
}
\arguments{
\item{X1}{training matrix to fit the quantile regression forest}

\item{Y1}{training vector}

\item{X2}{training matrix to compute the conformal scores}

\item{Y2}{training vector to compute the conformal scores}

\item{beta}{nominal quantile level}

\item{mtry}{random forest parameter}

\item{ntree}{random forest parameter}

\item{alpha}{miscoverage level}
}
\value{
a function for computing conditional width and coverage
}
\description{
Conditional width and coverage for CQR, internal function used inside conf_CQR_conditional
}
