% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_CQR_conditional.R
\name{conf_CQR_conditional}
\alias{conf_CQR_conditional}
\title{Conditional width and coverage for CQR}
\usage{
conf_CQR_conditional(x, y, beta, mtry, ntree, alpha = 0.1)
}
\arguments{
\item{x}{A N*d training matrix}

\item{y}{A N*1 training vector}

\item{beta}{nominal quantile level}

\item{mtry}{random forest parameter}

\item{ntree}{random forest parameter}

\item{alpha}{miscoverage level}
}
\value{
a function for computing conditional width and coverage
}
\description{
Conditional width and coverage for CQR
}
