% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_CQR_prelim.R
\name{conf_CQR_prelim}
\alias{conf_CQR_prelim}
\title{preliminary function for CQR}
\usage{
conf_CQR_prelim(X1, Y1, X2, Y2, beta_grid, mtry, ntree, alpha = 0.1)
}
\arguments{
\item{X1}{A n1*d matrix for training}

\item{Y1}{A n1*1 vector for training}

\item{X2}{A n2*d matrix for calibration}

\item{Y2}{A n2*1 vector for calibration}

\item{beta_grid}{a grid of beta's}

\item{mtry}{mtry parameter in random forest}

\item{ntree}{number of trees parameter in random forest}

\item{alpha}{miscoverage level}
}
\value{
the smallest width and its corresponding beta
}
\description{
preliminary function for CQR
}
