% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_CQR_reg.R
\name{conf_CQR_reg}
\alias{conf_CQR_reg}
\title{EFCP and VFCP for CQR, CQR-m, CQR-r}
\usage{
conf_CQR_reg(
  x,
  y,
  split,
  beta_grid,
  mtry_grid,
  ntree_grid,
  method = "efficient",
  alpha = 0.1
)
}
\arguments{
\item{x}{A N*d training matrix}

\item{y}{A N*1 training vector}

\item{split}{a vector of length 1 for efcp, length 2 for vfcp}

\item{beta_grid}{a grid of beta's}

\item{mtry_grid}{a grid of mtry}

\item{ntree_grid}{a grid of ntree}

\item{method}{"efficient" for efcp; "valid" for vfcp}

\item{alpha}{miscoverage level}
}
\value{
the selected cqr method
}
\description{
EFCP and VFCP for CQR, CQR-m, CQR-r
}
