% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_funs.R
\name{cv.fun}
\alias{cv.fun}
\title{Cross validation conformal prediction for ridge regression}
\usage{
cv.fun(X, Y, X0, lambda = seq(0, 100, length = 100), nfolds = 10, alpha = 0.1)
}
\arguments{
\item{X}{A N*d training matrix}

\item{Y}{A N*1 training vector}

\item{X0}{A N0*d testing vector}

\item{lambda}{a sequence of penalty parameters for ridge regression}

\item{nfolds}{number of folds}

\item{alpha}{miscoverage level}
}
\value{
upper and lower prediction intervals for X0
}
\description{
Cross validation conformal prediction for ridge regression
}
