% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_funs.R
\name{ginverselm.funs}
\alias{ginverselm.funs}
\title{Internal function used for ginverse.fun}
\usage{
ginverselm.funs(intercept = TRUE, lambda = 0)
}
\arguments{
\item{intercept}{default is TRUE}

\item{lambda}{a vector}
}
\description{
Internal function used for ginverse.fun
}
