% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_funs.R
\docType{data}
\name{my.ginverselm.funs}
\alias{my.ginverselm.funs}
\title{Internal function used for ginverse.fun}
\format{
An object of class \code{list} of length 4.
}
\usage{
my.ginverselm.funs
}
\description{
Internal function used for ginverse.fun
}
\keyword{datasets}
