% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_funs.R
\name{naive.fun}
\alias{naive.fun}
\title{Conformal prediction for linear regression}
\usage{
naive.fun(X, Y, X0, alpha = 0.1)
}
\arguments{
\item{X}{A N*d training matrix}

\item{Y}{A N*1 training vector}

\item{X0}{A N0*d testing vector}

\item{alpha}{miscoverage level}
}
\value{
upper and lower prediction intervals for X0
}
\description{
Conformal prediction for linear regression
}
