% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ridge_linear_cov100_t3}
\alias{ridge_linear_cov100_t3}
\title{Outcomes of an example for tuning-free conformal prediction with ridge regression.}
\format{
A list with 7 elements: dim_linear_t3,cov.param_linear_fm_t3, cov.naive_linear_fm_t3, cov.vfcp_linear_fm_t3, cov.star_linear_fm_t3, cov.cv5_linear_fm_t3, cov.efcp_linear_fm_t3
\describe{
\item{dim}{dimensions used in the experiment}
\item{len.param}{a matrix with coverages for the prediction regions produced by the parametric method}
\item{len.naive}{a matrix with coverages for the prediction regions produced by naive linear regression method}
\item{len.vfcp}{na matrix with coverages for the prediction regions produced by VFCP}
\item{len.star}{a matrix with coverages for the prediction regions produced by cross validation with the errors}
\item{len.cv5}{a matrix with coverages for the prediction regions produced by cross-validation with 5 splits}
\item{len.efcp}{a matrix with coverages for the prediction regions produced by efcp}
}
}
\source{
For details please see the "Example-tuning_free_ridge_regression" vignette:\code{vignette("Example-tuning_free_ridge_regression", package = "ConformalSmallest")}
}
\usage{
ridge_linear_cov100_t3
}
\description{
A dataset containing the experiment results used in the vignettes.
}
\keyword{datasets}
