% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ridge_linear_cov100_t5}
\alias{ridge_linear_cov100_t5}
\title{Outcomes of an example for tuning-free conformal prediction with ridge regression.}
\format{
A list with 7 elements: dim_linear_t5,cov.param_linear_fm_t5, cov.naive_linear_fm_t5, cov.vfcp_linear_fm_t5, cov.star_linear_fm_t5, cov.cv5_linear_fm_t5, cov.efcp_linear_fm_t5
\describe{
\item{dim}{dimensions used in the experiment}
\item{cov.param}{a matrix with coverages for the prediction regions produced by the parametric method}
\item{cov.naive}{a matrix with coverages for the prediction regions produced by naive linear regression method}
\item{cov.vfcp}{na matrix with coverages for the prediction regions produced by VFCP}
\item{cov.star}{a matrix with coverages for the prediction regions produced by cross validation with the errors}
\item{cov.cv5}{a matrix with coverages for the prediction regions produced by cross-validation with 5 splits}
\item{cov.efcp}{a matrix with coverages for the prediction regions produced by efcp}
}
}
\source{
For details please see the "Example-tuning_free_ridge_regression" vignette:\code{vignette("Example-tuning_free_ridge_regression", package = "ConformalSmallest")}
}
\usage{
ridge_linear_cov100_t5
}
\description{
A dataset containing the experiment results used in the vignettes.
}
\keyword{datasets}
