% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ridge_linear_len100_t5}
\alias{ridge_linear_len100_t5}
\title{Outcomes of an example for tuning-free conformal prediction with ridge regression.}
\format{
A list with 6 elements: len.param_linear_fm_t5, len.naive_linear_fm_t5, len.vfcp_linear_fm_t5, len.star_linear_fm_t5, len.cv5_linear_fm_t5, len.efcp_linear_fm_t5
\describe{
\item{len.param}{a matrix with widths for the prediction regions produced by the parametric method}
\item{len.naive}{a matrix with widths for the prediction regions produced by naive linear regression method}
\item{len.vfcp}{na matrix with widths for the prediction regions produced by VFCP}
\item{len.star}{a matrix with widths for the prediction regions produced by cross validation with the errors}
\item{len.cv5}{a matrix with widths for the prediction regions produced by cross-validation with 5 splits}
\item{len.efcp}{a matrix with widths for the prediction regions produced by efcp}
}
}
\source{
For details please see the "Example-tuning_free_ridge_regression" vignette:\code{vignette("Example-tuning_free_ridge_regression", package = "ConformalSmallest")}
}
\usage{
ridge_linear_len100_t5
}
\description{
A dataset containing the experiment results used in the vignettes.
}
\keyword{datasets}
