% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{superconduct}
\alias{superconduct}
\title{Superconduct data}
\format{
A dataset of dimension 81+1
}
\source{
train.csv
}
\usage{
superconduct
}
\description{
Superconduct data
}
\keyword{datasets}
