% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfcp_cqr.R
\name{vfcp_cqr}
\alias{vfcp_cqr}
\title{Validity first conformal prediction for Conformal Quantile Regression}
\usage{
vfcp_cqr(x, y, split, beta_grid, params_grid, alpha = 0.1)
}
\arguments{
\item{x}{A N*d training matrix}

\item{y}{A N*1 training vector}

\item{split}{a number between 0 and 1}

\item{beta_grid}{a grid of beta's}

\item{params_grid}{a grid of mtry and ntree}

\item{alpha}{miscoverage level}
}
\value{
average prediction width and a function for coverage on some testing points
}
\description{
Validity first conformal prediction for Conformal Quantile Regression
}
